problem using web browser

+++ src/tritonhttp/httpd_structs.go
@@ -1,25 +1,24 @@
 package tritonhttp
 
-type HttpServer	struct {
-	ServerPort	string
-	DocRoot		string
-	MIMEPath	string
-	MIMEMap		map[string]string
+type HttpServer struct {
+	ServerPort string
+	DocRoot    string
+	MIMEPath   string
+	MIMEMap    map[string]string
 }
 
 type HttpResponseHeader struct {
 	// Add any fields required for the response here
-	server string
-	lastModified string
-	contentType string
-	contentLength int
-	connection string
-
+	server        string
+	lastModified  string
+	contentType   string
+	contentLength int64
+	connection    string
 }
 
 type HttpRequestHeader struct {
 	// Add any fields required for the request here
-	host string
+	host       string
 	connection string
-	others bool
+	others     bool
 }
